240416 : mainPageController dy,hj 를 product, customer 컨트롤러로 나눔
