+++ ec21/src/main/resources/templates/main/productsWrite.html
@@ -107,7 +107,7 @@
 					<div class="col-lg-2 col-md-2 col-12">
 						<!-- Logo -->
 						<div class="logo">
-							<a th:href="@{/main/index}"><img th:src="@{/main/images/ec21logo.png}" alt="logo"></a>
+							<a th:href="@{/}"><img th:src="@{/main/images/ec21logo.png}" alt="logo"></a>
 						</div>
 						<!--/ End Logo -->
 						<!-- Search Form -->
@@ -136,10 +136,10 @@
 				<div class="col-12">
 					<div class="bread-inner">
 						<ul class="bread-list">
-							<li><a th:href="@{/main/index}">Home<i class="ti-arrow-right"></i></a></li>
+							<li><a th:href="@{/}">Home<i class="ti-arrow-right"></i></a></li>
 							<li class="active"><a href="#">Products</a></li>
 							<li><i class="ti-arrow-right"></i></a></li>
-							<li class="active"><a th:href="@{/main/[productsWrite]}">Add a New Product</a></li>
+							<li class="active"><a th:href="@{/main/productsWrite}">Add a New Product</a></li>
 						</ul>
 					</div>
 				</div>
@@ -357,7 +357,7 @@
 							</div>
 							<div class="col-lg-6 col-12">
 								<div class="right">
-									<img th:src="@{images/payments.png}" alt="#">
+									<img th:src="@{/images/payments.png}" alt="#">
 								</div>
 							</div>
 						</div>
@@ -375,7 +375,7 @@
 		<!-- Popper JS -->
 		<script th:src="@{/main/js/popper.min.js}"></script>
 		<!-- Bootstrap JS -->
-		<script th:src="@{static/main/js/bootstrap.min.js}"></script>
+		<script th:src="@{/static/main/js/bootstrap.min.js}"></script>
 		<!-- Color JS -->
 		<script th:src="@{/main/js/colors.js}"></script>
 		<!-- Slicknav JS -->
